feat(payroll): FnF settlement, fractional LOP, salary-revision arrears (WIP continuation)
Continues the salary-revision-arrears work left uncommitted in the working tree:
- FnF: UpdateFnfRun/GetFnfPreview/GetFnfSummary + orchestrator + settlement page.
- LOP: fractional LOP days (SetLop/BulkImportLop + FractionalLopDays migration).
- Exits: InitiateExit + exit YTD snapshot columns.
- Exports: payroll details / bank advice export adjustments.
- Migration consolidation: replace the Jun-02 migrations with
  RestoreFnfExitDocsSalaryRevision; delete the orphaned AddExitYtdSnapshot.cs
  (its Designer was removed and its ytd_*_snapshot columns are recreated by the
  consolidation migration). ModelSnapshot updated to match.
- Tests + web pages (FnF breakdown, import modal, set-password, provision org).
- Add .config/dotnet-tools.json (EF tool manifest).

+++ src/Payroll.Api/Controllers/PayrollRunsController.cs
@@ -510,7 +510,7 @@ public sealed class PayrollRunsController(ISender sender, ITenantContext tenantC
 }
 
 // Request body records
-public record SetLopRequest(int LopDays);
+public record SetLopRequest(decimal LopDays);
 public record AddOneTimeEarningRequest(Guid ComponentId, decimal Amount);
 public record OverrideTdsRequest(decimal OverrideAmount, string? Reason);
 public record SkipEmployeeRequest(string Reason);
@@ -519,7 +519,7 @@ public record RecordPaymentRequest(DateOnly PaymentDate, string PaymentMode, str
 public record UpdateSettlementDateRequest(DateOnly SettlementDate);
 
 public record UpdateFnfRunRequest(
-    int LopDays,
+    decimal LopDays,
     decimal Bonus,
     decimal Commission,
     decimal LeaveEncashment,
